error pages refactored and craco removed
