+++ packages/ui-components/src/error-pages/ErrorPage.ts
@@ -1,0 +1,11 @@
+import Page401 from './page-401/Page401'
+import Page404 from './page-404/Page404'
+import Page500 from './page-500/Page500'
+
+const PlaceholdersPage = {
+  Error401: Page401,
+  Error404: Page404,
+  Error500: Page500
+}
+
+export default PlaceholdersPage
