+++ packages/ui-components/src/error-pages/page-404/Page404.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import '../styles/errorPages.css'
+import '../../styles/errorPages.css'
 
 const Page404: React.FC = () => {
   return (
