+++ packages/ui-components/src/index.ts
@@ -1,4 +1,2 @@
 export {default as LoadingAnimation} from './loading-animation/LoadingAnimation'
-export {default as Page401} from './page-401/Page401'
-export {default as Page404} from './page-404/Page404'
-export {default as Page500} from './page-500/Page500'
+export {default as PlaceholdersPage} from './error-pages/ErrorPage'
