+++ packages/ui-components/src/styles/errorPages.css
@@ -1,5 +1,5 @@
 @import "./_colors";
 
-svg > .svg_primaryColor {
+.svg_primaryColor {
   fill: var(--ui-components-primary-color-tint-89);
 }
